requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==21.2.0
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
